chore: rename command to bearer jwt
