+++ internal/gen/bearerjwt/bearerjwt.go
@@ -1,4 +1,4 @@
-package token
+package bearerjwt
 
 import (
 	"context"
